+++ DawnLib.Dusk/src/API/Definitions/EntityReplacement/MapObjects/DuskMapObject.cs
@@ -5,7 +5,6 @@ using Dawn.Internal;
 using Dawn.Interfaces;
 using Dusk.Internal;
 using UnityEngine;
-using Dawn.Utils;
 
 namespace Dusk;