+++ DawnLib.Dusk/src/API/Definitions/EntityReplacement/Unlockables/DuskUnlockable.cs
@@ -6,7 +6,6 @@ using Dawn.Interfaces;
 using Dusk.Internal;
 using Newtonsoft.Json.Linq;
 using UnityEngine;
-using Dawn.Utils;
 
 namespace Dusk;