+++ DawnLib/src/API/DeadBodies/DawnDeadBodyNamespacedKeyContainer.cs
@@ -1,0 +1,13 @@
+using Dawn;
+using UnityEngine;
+
+public class DawnDeadBodyNamespacedKeyContainer : MonoBehaviour
+{
+    [field: SerializeField]
+    public NamespacedKey Value { get; internal set; }
+
+    public DawnDeadBodyInfo GetDawnDeadBodynfo()
+    {
+        return LethalContent.DeadBodies[Value.AsTyped<DawnDeadBodyInfo>()];
+    }
+}