+++ DawnLib/src/API/Enemies/DawnEnemyInfo.cs
@@ -4,33 +4,21 @@ namespace Dawn;
 
 public sealed class DawnEnemyInfo : DawnBaseInfo<DawnEnemyInfo>
 {
-    internal DawnEnemyInfo(NamespacedKey<DawnEnemyInfo> key, HashSet<NamespacedKey> tags, EnemyType enemyType, DawnEnemyLocationInfo? outside, DawnEnemyLocationInfo? inside, DawnEnemyLocationInfo? daytime, DawnEnemyLocationInfo? weed, TerminalNode? bestiaryNode, TerminalKeyword? nameKeyword, IDataContainer? customData) : base(key, tags, customData)
+    internal DawnEnemyInfo(NamespacedKey<DawnEnemyInfo> key, HashSet<NamespacedKey> tags, EnemyType enemyType, DawnEnemyLocationInfo outside, DawnEnemyLocationInfo inside, DawnEnemyLocationInfo daytime, DawnEnemyLocationInfo weed, TerminalNode? bestiaryNode, TerminalKeyword? nameKeyword, IDataContainer? customData) : base(key, tags, customData)
     {
         EnemyType = enemyType;
 
         Outside = outside;
-        if (Outside != null)
-        {
-            Outside.ParentInfo = this;
-        }
+        Outside.ParentInfo = this;
 
         Inside = inside;
-        if (Inside != null)
-        {
-            Inside.ParentInfo = this;
-        }
+        Inside.ParentInfo = this;
 
         Daytime = daytime;
-        if (Daytime != null)
-        {
-            Daytime.ParentInfo = this;
-        }
+        Daytime.ParentInfo = this;
 
         Weed = weed;
-        if (Weed != null)
-        {
-            Weed.ParentInfo = this;
-        }
+        Weed.ParentInfo = this;
 
         BestiaryNode = bestiaryNode;
         NameKeyword = nameKeyword;
@@ -38,10 +26,10 @@ public sealed class DawnEnemyInfo : DawnBaseInfo<DawnEnemyInfo>
 
     public EnemyType EnemyType { get; }
 
-    public DawnEnemyLocationInfo? Outside { get; private set; }
-    public DawnEnemyLocationInfo? Inside { get; private set; }
-    public DawnEnemyLocationInfo? Daytime { get; private set; }
-    public DawnEnemyLocationInfo? Weed { get; private set; }
+    public DawnEnemyLocationInfo Outside { get; private set; }
+    public DawnEnemyLocationInfo Inside { get; private set; }
+    public DawnEnemyLocationInfo Daytime { get; private set; }
+    public DawnEnemyLocationInfo Weed { get; private set; }
 
     public TerminalNode? BestiaryNode { get; internal set; }
     public TerminalKeyword? NameKeyword { get; internal set; }