+++ DawnLib/src/API/Surfaces/PlayerControllerBExtensions.cs
@@ -5,7 +5,7 @@ using GameNetcodeStuff;
 
 namespace Dawn;
 
-public static class PlayerControllerBExtensions
+public static partial class PlayerControllerBExtensions
 {
     public static bool TryGetCurrentDawnSurface(this PlayerControllerB player, [NotNullWhen(true)] out DawnSurface? dawnSurface)
     {